updated cases if output files are empty

+++ snakefiles/ALIGNMENT.snakefile
@@ -45,16 +45,20 @@ rule alignment:
                     score_writer.write("\n".join(scores))
 
                 del scores[:]
-            else:
+                
+            if(not os.path.exists(output[0])):
                 with open(output[0], "w") as empty_writer:
                     empty_writer.write("")
 
+            if(not os.path.exists(output[1])):
                 with open(output[1], "w") as empty_writer:
                     empty_writer.write("")
 
+            if(not os.path.exists(output[2])):
                 with open(output[2], "w") as empty_writer:
                     empty_writer.write("")
 
+            if(not os.path.exists(output[3])):
                 with open(output[3], "w") as empty_writer:
                     empty_writer.write("")
         except Exception as ex:
